UserTransactions Updated

+++ Recodme.Labs.MarketAnalyzer.MarketAnalyzerDB/dbo/Tables/UserTransactions.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[UserTransactions] (
     [Id]                      UNIQUEIDENTIFIER NOT NULL,
-    [NumberOfShares]          MONEY            NULL,
+    [NumberOfShares]          FLOAT (53)       NULL,
     [ValueOfShares]           MONEY            NULL,
-    [NumberOfSharesWithdrawn] MONEY            NULL,
+    [NumberOfSharesWithdrawn] FLOAT (53)       NULL,
     [ValueOfSharesWithdrawn]  MONEY            NULL,
     [DateOfMovement]          DATETIME2 (3)    NULL,
     [DateCreated]             DATETIME2 (3)    NULL,
@@ -13,6 +13,8 @@
     CONSTRAINT [FK__UserTrans__AspNe__6FE99F9F] FOREIGN KEY ([AspNetUserId]) REFERENCES [dbo].[AspNetUsers] ([Id]),
     CONSTRAINT [FK__UserTrans__Compa__70DDC3D8] FOREIGN KEY ([CompanyId]) REFERENCES [dbo].[Companies] ([Id])
 );
+
+
